refactor: enhance MarketDetailV2 with improved header and skeleton components
- Replaced the previous header implementation in MarketDetailV2 with a new MarketDetailHeader component for better modularity and maintainability.
- Introduced MarketDetailSkeleton for loading states, enhancing user experience during data fetching.
- Updated component imports and structure to support the new header and skeleton integration, streamlining the overall codebase.
