+++ packages/kit/src/views/Market/MarketDetailV2/components/index.ts
@@ -1,4 +1,4 @@
-export * from './MarketDetailHeader';
 export * from './SwapPanel';
 export * from './TokenPriceChart';
 export * from './TokenDetailHeader';
+export * from './MarketDetailHeader';
